requirements.txt
torchstat
ptflops
torchsummary
torchvision
timm
matplotlib
addict
